更新処理を分割
