+++ .gitignore
@@ -26,3 +26,9 @@ CHAPTER6/シーン/終了画面/__pycache__/
 CHAPTER6/背景/__pycache__/
 CHAPTER6/特殊効果/__pycache__/
 CHAPTER6/共通/__pycache__/
+
+CHAPTER6/main/更新/__pycache__/
+CHAPTER6/main/更新/移動/__pycache__/
+CHAPTER6/main/更新/衝突/__pycache__/
+CHAPTER6/main/更新/発射/__pycache__/
+CHAPTER6/main/更新/出現/__pycache__/
