+++ CHAPTER6/main/DB.py
@@ -10,6 +10,7 @@
 プレイ時間  = 0     # 難易度の変化，出現タイミングの物差し
 難易度      = 0     # 敵の強さ、出現頻度の物差し
 難易度間隔  = 1800  # この感覚でレベルUp・ボスキャラ登場
+難易度間隔  = 300  # この感覚でレベルUp・ボスキャラ登場
 得点        = 0     # 得点
 人数        = 0     # プレイヤーの人数
