+++ CHAPTER6/main/main.py
@@ -7,10 +7,7 @@ import pyxel
 import main.DB
 from   シーン        import *
 from   背景	         import 背景作成
-from   .FN更新1_移動 import 移動クラス
-from   .FN更新2_発射 import 発射クラス
-from   .FN更新3_衝突 import 衝突クラス
-from   .FN更新4_出現 import 出現クラス
+from   .更新         import 更新_移動, 更新_発射, 更新_衝突, 更新_出現
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
@@ -63,10 +60,10 @@ class ゲーム本体:
 		#┬
         #≫更新処理のオブジェクト・セットを作る(タプル型）
         main.DB.FN更新 = (
-            移動クラス() ,
-            発射クラス() ,
-            衝突クラス() ,
-            出現クラス() )
+            更新_移動() ,
+            更新_発射() ,
+            更新_衝突() ,
+            更新_出現() )
         #│
         #≫シーン処理のオブジェクト・セットを作る(辞書型）
         main.DB.FNシーン = {
