+++ CHAPTER6/main/更新/__init__.py
@@ -1,0 +1,7 @@
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃処理：更新
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+from .移動    import *
+from .衝突    import *
+from .発射    import *
+from .出現    import *
