+++ CHAPTER6/main/更新/出現/main.py
@@ -24,7 +24,7 @@ class 仕様:
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 出現クラス:
+class 本体:
     #┌───────────────────────────────────
     #│機能実行
     #└───────────────────────────────────
