+++ CHAPTER6/main/更新/発射/__init__.py
@@ -1,0 +1,4 @@
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃処理：更新：発射
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+from .main import 本体 as 更新_発射
