+++ CHAPTER6/main/更新/発射/main.py
@@ -6,21 +6,21 @@ import main.DB
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 発射クラス:
+class 本体:
     #┌───────────────────────────────────
     #│機能実行
     #└───────────────────────────────────
     def 実行(self):
 		#┬
         #●自機を更新(弾を発射)する
-        self.Fn発射(main.DB.obj自機)
+        self.Fnリスト処理(main.DB.obj自機)
         #│
         #●敵機を更新(弾を発射)する
-        self.Fn発射(main.DB.obj敵機)
+        self.Fnリスト処理(main.DB.obj敵機)
 		#┴ 
 	#────────────────────────────────────	
-    def Fn発射(self         ,
-             引数_対象一覧):  #① 対象オブジェクト一覧(自機/敵機) 
+    def Fnリスト処理(self   ,
+             引数_対象一覧  ):  #① 対象オブジェクト一覧(自機/敵機) 
 		#┬
         #○対象オブジェクトの中味を確認する
         if not 引数_対象一覧: return
