+++ CHAPTER6/main/更新/移動/main.py
@@ -6,14 +6,14 @@ import main.DB
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 移動クラス:
+class 本体:
     #┌───────────────────────────────────
     #│機能実行
     #└───────────────────────────────────
     def 実行(self):
 		#┬
         #●└┐オブジェクトを更新(移動)する
-        self.Fn移動((
+        self.Fnセット処理((
                 main.DB.obj背景       ,
                 main.DB.obj特殊効果   ,
                 main.DB.obj自機       ,
@@ -25,27 +25,27 @@ class 移動クラス:
                 main.DB.FNシーン[ main.DB.シーン ]))
 		#┴ 
 	#────────────────────────────────────	
-    def Fn移動(self         ,
-             引数_処理一覧):  #① 処理の一覧 ※一覧の入れ子OK
+    def Fnセット処理(self   ,
+             引数_対象セット):  #① 処理の一覧 ※一覧の入れ子OK
 		#┬
 		#◎└┐すべてのオブジェクトを描画する
-        for tmp処理 in 引数_処理一覧:
+        for tmp対象 in 引数_対象セット:
 			#│＼（すべての処理を終えた場合）
             #│ ↓
             #│ ▼繰り返し処理を抜ける
             #│
             #◇┐要素オブジェクトを描画する
-            if isinstance(tmp処理, list):
+            if isinstance(tmp対象, list):
             #　├┐（一覧オブジェクトの場合）
                 #↓
                 #◎└┐要素オブジェクトを描画する
-                for tmp要素 in tmp処理:
+                for tmp要素 in tmp対象:
                     #│＼（すべての処理を終えた場合）
                     #│ ↓
                     #│ ▼繰り返し処理を抜ける
                     #│
                     #●オブジェクトの中味を確認する
-                    if tmp処理 is None: continue
+                    if tmp対象 is None: continue
                     #│＼（すべての処理を終えた場合）
                     #│ ↓
                     #│ ▼次の要素オブジェクトを処理する
@@ -54,10 +54,10 @@ class 移動クラス:
                     tmp要素.FN移動.実行()
                     #┴ 
 
-            elif tmp処理 is not None:
+            elif tmp対象 is not None:
             #　├┐（要素で中味がある場合）
                 #↓
                 #●オブジェクトを移動する
-                tmp処理.FN移動.実行()
+                tmp対象.FN移動.実行()
             #　└┐（その他）
         #┴　┴　┴
