+++ CHAPTER6/main/更新/衝突/Fn2_衝突.py
@@ -1,0 +1,50 @@
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃更新機能：衝突処理
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+import main.DB
+
+#┌───────────────────────────────────
+#│衝突処理
+#└───────────────────────────────────
+def Fn衝突_攻撃(        #【戻り値】 True:破壊された／False：存在する
+        引数_弾     ,   #① 自分のオブジェクト(弾)
+        引数_対象   ):  #② 衝突対象のオブジェクト(アイテム,敵機,弾)
+    #┬
+    #○条件を確認する
+    if 引数_対象.仕様.アイテム区分: return
+    #│＼（アイテムの場合）
+    #│ ↓
+    #│ ▼処理を中断する
+    #│
+    #●攻撃対象にダメージを与える
+    #●自分の弾を破壊する
+    引数_対象.FN衝突.実行( main.DB.obj自機共通.仕様.威力_弾 )
+    引数_弾.FN衝突.実行()
+    #│
+    #▼弾の破壊可否を返す
+    return (引数_弾 not in main.DB.obj弾_自機)
+
+#────────────────────────────────────	
+def Fn衝突_体当たり(        #【戻り値】 True:破壊された／False：存在する
+        引数_自機       ,   #① 自分のオブジェクト(自機)
+        引数_対象       ):  #② 衝突対象のオブジェクト(アイテム,敵機,弾)
+    #┬
+    #◇┐衝突を処理する
+    if 引数_対象.仕様.アイテム区分:
+    #　├┐（相手が『アイテム』の場合）
+        #↓
+        #●アイテムの効果を発動する
+        #●アイテムを消滅する
+        main.DB.obj特殊効果.FN衝突.実行( 引数_対象 )
+        引数_対象.FN衝突.実行( 0, 0 )
+        #┴ 
+    else:
+    #　└┐（その他）
+        #↓
+        #●自機にダメージ(体当たり)を与える
+        #●相手にダメージ(体当たり)を与える
+        引数_自機.FN衝突.実行( 引数_対象.FN衝突.仕様.威力 )
+        引数_対象.FN衝突.実行( main.DB.obj自機共通.仕様.威力_体当たり )
+    #│
+    #▼自機の破壊状況を返す
+    return (引数_自機 not in main.DB.obj自機)
