+++ CHAPTER6/main/更新/衝突/main.py
@@ -1,0 +1,25 @@
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃更新機能：衝突
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+from  .Fn0_セット処理 import *
+
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃メイン
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+class 本体:
+    #┌───────────────────────────────────
+    #│機能実行
+    #└───────────────────────────────────
+    def 実行(self):
+		#┬
+        #○自機の弾攻撃で、更新(衝突判定)する
+        Fnセット処理_攻撃( [
+                main.DB.obj敵機     ,
+                main.DB.obj弾_敵機  ] )
+        #│
+        #○自機の体当たりで、更新(衝突判定)する
+        Fnセット処理_体当たり( [
+                main.DB.objアイテム ,
+                main.DB.obj敵機     ,
+                main.DB.obj弾_敵機  ] )
+        #┴
